+++ modules/distribution-zones/src/main/java/org/apache/ignite/internal/distributionzones/events/HaZoneTopologyUpdateEvent.java
@@ -1,0 +1,32 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.distributionzones.events;
+
+import org.apache.ignite.internal.event.Event;
+
+/**
+ * Events produced by the {@link org.apache.ignite.internal.distributionzones.DistributionZoneManager}
+ * about the changes in topology of zones with
+ * {@link org.apache.ignite.internal.catalog.descriptors.ConsistencyMode#HIGH_AVAILABILITY} consistency mode.
+ */
+// TODO: https://issues.apache.org/jira/browse/IGNITE-23640 Remove this event and
+//  replace it by the direct call to the disaster recovery manager.
+@Deprecated
+public enum HaZoneTopologyUpdateEvent implements Event {
+    TOPOLOGY_REDUCED
+}
