--- modules/distribution-zones/src/main/java/org/apache/ignite/internal/distributionzones/events/HighAvalabilityZoneTopologyUpdateEvent.java
@@ -1,7 +1,0 @@
-package org.apache.ignite.internal.distributionzones.events;
-
-import org.apache.ignite.internal.event.Event;
-
-public enum HighAvalabilityZoneTopologyUpdateEvent implements Event {
-    TOPOLOGY_REDUCED
-}
