+++ modules/metastorage-api/src/main/java/org/apache/ignite/internal/metastorage/MetaStorageManager.java
@@ -273,8 +273,6 @@ public interface MetaStorageManager extends IgniteComponent {
      */
     HybridTimestamp timestampByRevisionLocally(long revision);
 
-    long revisionByTimestamoLocally(HybridTimestamp timestamp);
-
     /**
      * Returns a future of getting the latest version of entries corresponding to the given keys from the metastorage leader.
      *
