+++ modules/metastorage/src/main/java/org/apache/ignite/internal/metastorage/impl/MetaStorageManagerImpl.java
@@ -888,11 +888,6 @@ public class MetaStorageManagerImpl implements MetaStorageManager, MetastorageGr
     }
 
     @Override
-    public long revisionByTimestamoLocally(HybridTimestamp timestamp) {
-        return inBusyLock(busyLock, () -> storage.revisionByTimestamp(timestamp));
-    }
-
-    @Override
     public CompletableFuture<Map<ByteArray, Entry>> getAll(Set<ByteArray> keys) {
         return inBusyLock(
                 busyLock,
