+++ modules/table/src/integrationTest/java/org/apache/ignite/internal/table/distributed/disaster/ItHighAvailablePartitionsRecoveryTest.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.apache.ignite.internal.table.distributed.disaster;
 
 import static org.apache.ignite.internal.catalog.CatalogService.DEFAULT_STORAGE_PROFILE;
@@ -21,6 +38,7 @@ import org.apache.ignite.internal.versioned.VersionedSerialization;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+/** Test for the HA zones recovery. */
 public class ItHighAvailablePartitionsRecoveryTest  extends ClusterPerTestIntegrationTest {
     private static String ZONE_NAME = "HA_ZONE";
 
@@ -77,9 +95,7 @@ public class ItHighAvailablePartitionsRecoveryTest  extends ClusterPerTestIntegr
 
         assertEquals(zoneId, request.zoneId());
         assertEquals(tableId, request.tableId());
-        assertEquals(Set.of(0,1), request.partitionIds());
+        assertEquals(Set.of(0, 1), request.partitionIds());
         assertFalse(request.manualUpdate());
     }
-
-
 }
